requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
openai>=1.10.0

# Vector store
chromadb>=0.4.0

# Document loading
pypdf>=3.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
